requirements.txt
asgiref==3.4.1
crispy-tailwind==0.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.5
django-crispy-forms==1.12.0
django-filter==2.4.0
django-heroku==0.3.1
django-phonenumber-field==5.2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
gunicorn==20.1.0
phonenumbers==8.12.28
Pillow==8.3.1
psycopg2==2.9.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
whitenoise==5.3.0
